+++ Source/FDChessboardView.swift
@@ -20,7 +20,7 @@ public struct FDChessboardSquare {
     /// A format like a4
     public var algebriac: String {
         get {
-            return String(UnicodeScalar(96 + file)) + String(rank + 1)
+            return String(describing: UnicodeScalar(96 + file)) + String(rank + 1)
         }
     }
 
@@ -81,86 +81,86 @@ public enum FDChessboardPiece: String {
 
 public protocol FDChessboardViewDataSource: class {
     /// What piece is on the square?
-    func chessboardView(board: FDChessboardView, pieceForSquare square: FDChessboardSquare) -> FDChessboardPiece?
+    func chessboardView(_ board: FDChessboardView, pieceForSquare square: FDChessboardSquare) -> FDChessboardPiece?
 
     /// The last move
-    func chessboardViewLastMove(board: FDChessboardView) -> (from:FDChessboardSquare, to:FDChessboardSquare)?
+    func chessboardViewLastMove(_ board: FDChessboardView) -> (from:FDChessboardSquare, to:FDChessboardSquare)?
 
     /// The premove
-    func chessboardViewPremove(board: FDChessboardView) -> (from:FDChessboardSquare, to:FDChessboardSquare)?
+    func chessboardViewPremove(_ board: FDChessboardView) -> (from:FDChessboardSquare, to:FDChessboardSquare)?
 }
 
 public protocol FDChessboardViewDelegate: class {
     /// Where can this piece move to?
-    func chessboardView(board: FDChessboardView, legalDestinationsForPieceAtSquare from: FDChessboardSquare) -> [FDChessboardSquare]
+    func chessboardView(_ board: FDChessboardView, legalDestinationsForPieceAtSquare from: FDChessboardSquare) -> [FDChessboardSquare]
 
     /// Before a move happens (cannot be stopped)
-    func chessboardView(board: FDChessboardView, willMoveFrom from: FDChessboardSquare, to: FDChessboardSquare)
+    func chessboardView(_ board: FDChessboardView, willMoveFrom from: FDChessboardSquare, to: FDChessboardSquare)
 
     /// After a move happened
-    func chessboardView(board: FDChessboardView, didMoveFrom from: FDChessboardSquare, to: FDChessboardSquare)
+    func chessboardView(_ board: FDChessboardView, didMoveFrom from: FDChessboardSquare, to: FDChessboardSquare)
 
     /// Before a move happens (cannot be stopped)
-    func chessboardView(board: FDChessboardView, willMoveFrom from: FDChessboardSquare, to: FDChessboardSquare, withPromotion promotion: FDChessboardPiece)
+    func chessboardView(_ board: FDChessboardView, willMoveFrom from: FDChessboardSquare, to: FDChessboardSquare, withPromotion promotion: FDChessboardPiece)
 
     /// After a move happened
-    func chessboardView(board: FDChessboardView, didMoveFrom from: FDChessboardSquare, to: FDChessboardSquare, withPromotion promotion: FDChessboardPiece)
+    func chessboardView(_ board: FDChessboardView, didMoveFrom from: FDChessboardSquare, to: FDChessboardSquare, withPromotion promotion: FDChessboardPiece)
 }
 
 /// Display for a chess board
-@IBDesignable public class FDChessboardView: UIView {
+@IBDesignable open class FDChessboardView: UIView {
     /// Color for "white" board squares
-    @IBInspectable public var lightBackgroundColor: UIColor = UIColor(red: 222.0/255, green:196.0/255, blue:160.0/255, alpha:1)
+    @IBInspectable open var lightBackgroundColor: UIColor = UIColor(red: 222.0/255, green:196.0/255, blue:160.0/255, alpha:1)
 
     /// Color for "black" board squares
-    @IBInspectable public var darkBackgroundColor: UIColor = UIColor(red: 160.0/255, green:120.0/255, blue:55.0/255, alpha:1)
+    @IBInspectable open var darkBackgroundColor: UIColor = UIColor(red: 160.0/255, green:120.0/255, blue:55.0/255, alpha:1)
 
     /// Color for where a piece is moving to
-    public var targetBackgroundColor = UIColor(hue: 0.75, saturation:0.5, brightness:0.5, alpha:1.0)
+    open var targetBackgroundColor = UIColor(hue: 0.75, saturation:0.5, brightness:0.5, alpha:1.0)
 
     /// Color for a legal square target
-    public var legalBackgroundColor = UIColor(hue: 0.25, saturation:0.5, brightness:0.5, alpha:1.0)
+    open var legalBackgroundColor = UIColor(hue: 0.25, saturation:0.5, brightness:0.5, alpha:1.0)
 
     /// Color for the last move square
-    public var lastMoveColor = UIColor(hue: 0.35, saturation:0.5, brightness:0.5, alpha:1.0)
+    open var lastMoveColor = UIColor(hue: 0.35, saturation:0.5, brightness:0.5, alpha:1.0)
 
     /// Color for a premove square
-    public var premoveColor = UIColor(hue: 0.15, saturation:0.5, brightness:0.5, alpha:1.0)
+    open var premoveColor = UIColor(hue: 0.15, saturation:0.5, brightness:0.5, alpha:1.0)
 
     /// Path for custom piece graphics
-    public var pieceGraphicsDirectoryPath: String? = nil
+    open var pieceGraphicsDirectoryPath: String? = nil
 
     /// Datasource to say which pieces are on each square
-    public weak var dataSource: FDChessboardViewDataSource? = nil {
+    open weak var dataSource: FDChessboardViewDataSource? = nil {
         didSet {
             reloadData()
         }
     }
 
     /// Handler for user interaction with the view
-    public weak var delegate: FDChessboardViewDelegate? = nil
+    open weak var delegate: FDChessboardViewDelegate? = nil
 
     /// Should piece moves be animated?
-    public var doesAnimate: Bool = true
+    open var doesAnimate: Bool = true
 
     /// Should legal squares be shown when a piece is selected?
-    public var doesShowLegalSquares: Bool = true
+    open var doesShowLegalSquares: Bool = true
 
     /// Should the lash move be shown?
-    public var doesShowLastMove: Bool = true
+    open var doesShowLastMove: Bool = true
 
     /// Should premove be shown?
-    public var doesShowPremove: Bool = true
+    open var doesShowPremove: Bool = true
 
-    private lazy var tilesAtIndex = [UIView]()
+    fileprivate lazy var tilesAtIndex = [UIView]()
 
-    private var pieceAtIndex = [Int : FDChessboardPiece]()
+    fileprivate var pieceAtIndex = [Int : FDChessboardPiece]()
 
-    private var pieceImageViewAtIndex = [Int : UIImageView]()
+    fileprivate var pieceImageViewAtIndex = [Int : UIImageView]()
 
-    private var lastMoveArrow: UIView? = nil
+    fileprivate var lastMoveArrow: UIView? = nil
 
-    private var premoveArrow: UIView? = nil
+    fileprivate var premoveArrow: UIView? = nil
 
     /// UIView initializer
     public override init(frame: CGRect) {
@@ -174,7 +174,7 @@ public protocol FDChessboardViewDelegate: class {
         setup()
     }
 
-    private func setup() {
+    fileprivate func setup() {
         self.translatesAutoresizingMaskIntoConstraints = false
 
         for index in 0..<64 {
@@ -187,25 +187,25 @@ public protocol FDChessboardViewDelegate: class {
 
             switch square.rank {
             case 0:
-                addConstraint(NSLayoutConstraint(item: tile, attribute: .Bottom, relatedBy: .Equal, toItem: self, attribute: .Bottom, multiplier: 1, constant: 0))
+                addConstraint(NSLayoutConstraint(item: tile, attribute: .bottom, relatedBy: .equal, toItem: self, attribute: .bottom, multiplier: 1, constant: 0))
             case 7:
-                addConstraint(NSLayoutConstraint(item: tile, attribute: .Bottom, relatedBy: .Equal, toItem: tilesAtIndex[index - 8], attribute: .Top, multiplier: 1, constant: 0))
-                addConstraint(NSLayoutConstraint(item: tile, attribute: .Top, relatedBy: .Equal, toItem: self, attribute: .Top, multiplier: 1, constant: 0))
-                addConstraint(NSLayoutConstraint(item: tile, attribute: .Height, relatedBy: .Equal, toItem: tilesAtIndex[index - 8], attribute: .Height, multiplier: 1, constant: 0))
+                addConstraint(NSLayoutConstraint(item: tile, attribute: .bottom, relatedBy: .equal, toItem: tilesAtIndex[index - 8], attribute: .top, multiplier: 1, constant: 0))
+                addConstraint(NSLayoutConstraint(item: tile, attribute: .top, relatedBy: .equal, toItem: self, attribute: .top, multiplier: 1, constant: 0))
+                addConstraint(NSLayoutConstraint(item: tile, attribute: .height, relatedBy: .equal, toItem: tilesAtIndex[index - 8], attribute: .height, multiplier: 1, constant: 0))
             default:
-                addConstraint(NSLayoutConstraint(item: tile, attribute: .Bottom, relatedBy: .Equal, toItem: tilesAtIndex[index - 8], attribute: .Top, multiplier: 1, constant: 0))
-                addConstraint(NSLayoutConstraint(item: tile, attribute: .Height, relatedBy: .Equal, toItem: tilesAtIndex[index - 8], attribute: .Height, multiplier: 1, constant: 0))
+                addConstraint(NSLayoutConstraint(item: tile, attribute: .bottom, relatedBy: .equal, toItem: tilesAtIndex[index - 8], attribute: .top, multiplier: 1, constant: 0))
+                addConstraint(NSLayoutConstraint(item: tile, attribute: .height, relatedBy: .equal, toItem: tilesAtIndex[index - 8], attribute: .height, multiplier: 1, constant: 0))
             }
             switch square.file {
             case 0:
-                addConstraint(NSLayoutConstraint(item: tile, attribute: .Left, relatedBy: .Equal, toItem: self, attribute: .Left, multiplier: 1, constant: 0))
+                addConstraint(NSLayoutConstraint(item: tile, attribute: .left, relatedBy: .equal, toItem: self, attribute: .left, multiplier: 1, constant: 0))
             case 7:
-                addConstraint(NSLayoutConstraint(item: tile, attribute: .Left, relatedBy: .Equal, toItem: tilesAtIndex[index - 1], attribute: .Right, multiplier: 1, constant: 0))
-                addConstraint(NSLayoutConstraint(item: tile, attribute: .Right, relatedBy: .Equal, toItem: self, attribute: .Right, multiplier: 1, constant: 0))
-                addConstraint(NSLayoutConstraint(item: tile, attribute: .Width, relatedBy: .Equal, toItem: tilesAtIndex[index - 1], attribute: .Width, multiplier: 1, constant: 0))
+                addConstraint(NSLayoutConstraint(item: tile, attribute: .left, relatedBy: .equal, toItem: tilesAtIndex[index - 1], attribute: .right, multiplier: 1, constant: 0))
+                addConstraint(NSLayoutConstraint(item: tile, attribute: .right, relatedBy: .equal, toItem: self, attribute: .right, multiplier: 1, constant: 0))
+                addConstraint(NSLayoutConstraint(item: tile, attribute: .width, relatedBy: .equal, toItem: tilesAtIndex[index - 1], attribute: .width, multiplier: 1, constant: 0))
             default:
-                addConstraint(NSLayoutConstraint(item: tile, attribute: .Left, relatedBy: .Equal, toItem: tilesAtIndex[index - 1], attribute: .Right, multiplier: 1, constant: 0))
-                addConstraint(NSLayoutConstraint(item: tile, attribute: .Width, relatedBy: .Equal, toItem: tilesAtIndex[index - 1], attribute: .Width, multiplier: 1, constant: 0))
+                addConstraint(NSLayoutConstraint(item: tile, attribute: .left, relatedBy: .equal, toItem: tilesAtIndex[index - 1], attribute: .right, multiplier: 1, constant: 0))
+                addConstraint(NSLayoutConstraint(item: tile, attribute: .width, relatedBy: .equal, toItem: tilesAtIndex[index - 1], attribute: .width, multiplier: 1, constant: 0))
             }
         }
         self.layoutIfNeeded()
@@ -236,34 +236,34 @@ tile.backgroundColor = self.legalBackgroundColor;
 */
 
     /// Add a piece onto the board
-    public func setPiece(piece: FDChessboardPiece?, forSquare square: FDChessboardSquare){
+    open func setPiece(_ piece: FDChessboardPiece?, forSquare square: FDChessboardSquare){
         let index = square.index
         pieceAtIndex[index] = piece
         self.pieceImageViewAtIndex[index]?.removeFromSuperview()
-        self.pieceImageViewAtIndex.removeValueForKey(index)
+        self.pieceImageViewAtIndex.removeValue(forKey: index)
         guard let piece = piece else {
             return
         }
 
         let pieceImageView: UIImageView
         let fileName = piece.rawValue
-        let bundle = NSBundle(forClass: self.dynamicType)
-        let image = UIImage(named: fileName, inBundle: bundle, compatibleWithTraitCollection: nil)
+        let bundle = Bundle(for: type(of: self))
+        let image = UIImage(named: fileName, in: bundle, compatibleWith: nil)
         pieceImageView = UIImageView(image: image)
         pieceImageView.translatesAutoresizingMaskIntoConstraints = false
         self.pieceImageViewAtIndex[index] = pieceImageView
         self.addSubview(pieceImageView)
 
         let squareOne = self.tilesAtIndex.first!
-        self.addConstraint(NSLayoutConstraint(item: pieceImageView, attribute: .Width, relatedBy: .Equal, toItem: squareOne, attribute: .Width, multiplier: 1, constant: 0))
-        self.addConstraint(NSLayoutConstraint(item: pieceImageView, attribute: .Height, relatedBy: .Equal, toItem: squareOne, attribute: .Height, multiplier: 1, constant: 0))
-        self.addConstraint(NSLayoutConstraint(item: pieceImageView, attribute: .Top, relatedBy: .Equal, toItem: self.tilesAtIndex[index], attribute: .Top, multiplier: 1, constant: 0))
-        self.addConstraint(NSLayoutConstraint(item: pieceImageView, attribute: .Leading, relatedBy: .Equal, toItem: self.tilesAtIndex[index], attribute: .Leading, multiplier: 1, constant: 0))
+        self.addConstraint(NSLayoutConstraint(item: pieceImageView, attribute: .width, relatedBy: .equal, toItem: squareOne, attribute: .width, multiplier: 1, constant: 0))
+        self.addConstraint(NSLayoutConstraint(item: pieceImageView, attribute: .height, relatedBy: .equal, toItem: squareOne, attribute: .height, multiplier: 1, constant: 0))
+        self.addConstraint(NSLayoutConstraint(item: pieceImageView, attribute: .top, relatedBy: .equal, toItem: self.tilesAtIndex[index], attribute: .top, multiplier: 1, constant: 0))
+        self.addConstraint(NSLayoutConstraint(item: pieceImageView, attribute: .leading, relatedBy: .equal, toItem: self.tilesAtIndex[index], attribute: .leading, multiplier: 1, constant: 0))
         self.layoutIfNeeded()
     }
 
     /// Repull all board information from data source
-    public func reloadData() {
+    open func reloadData() {
         for index in 0 ..< 64 {
             let square = FDChessboardSquare(index: index)
             let newPiece = dataSource?.chessboardView(self, pieceForSquare:square)
@@ -273,22 +273,22 @@ tile.backgroundColor = self.legalBackgroundColor;
     }
 
     /// Move a piece on the board
-    public func movePieceAtCoordinate(from: FDChessboardSquare, toCoordinate to: FDChessboardSquare) -> Bool {
+    open func movePieceAtCoordinate(_ from: FDChessboardSquare, toCoordinate to: FDChessboardSquare) -> Bool {
         return true
     }
 
     /// Move a piece on the board
-    public func movePieceAtCoordinate(from: FDChessboardSquare, toCoordinate to: FDChessboardSquare, andPromoteTo piece: FDChessboardPiece) -> Bool {
+    open func movePieceAtCoordinate(_ from: FDChessboardSquare, toCoordinate to: FDChessboardSquare, andPromoteTo piece: FDChessboardPiece) -> Bool {
         return true
     }
 
     /// Premove a piece on the board
-    public func premovePieceAtCoordinate(from: FDChessboardSquare, toCoordinate to: FDChessboardSquare) -> Bool {
+    open func premovePieceAtCoordinate(_ from: FDChessboardSquare, toCoordinate to: FDChessboardSquare) -> Bool {
         return true
     }
 
     /// Premove a piece on the board
-    public func premovePieceAtCoordinate(from: FDChessboardSquare, toCoordinate to: FDChessboardSquare, andPromoteTo piece: FDChessboardPiece) -> Bool {
+    open func premovePieceAtCoordinate(_ from: FDChessboardSquare, toCoordinate to: FDChessboardSquare, andPromoteTo piece: FDChessboardPiece) -> Bool {
         return true
     }